eyeos-modal-window is now eyeos-modal-window-lib
We seem to have lost the mapping between repo name and bower package
same, so they are now both the same. Instead of fixing the mapping, I'm
going with the easier option of just renaming the bower.json.

This is not something I care a lot about since we will be moving away
from bower fairly soon.