Create logoutNotifier addon/module
This notifies you that you are no longer in a valid session when you
have multiple open365 tabs open and then you logout from the platform in
one of them. When you visit any of the open365 tabs that had your
session opened it will pop the notification with a button to log in
again.